Add files via upload
This is a working example of code for Fig2 Single-cell correlates with task variables(visual stimulus). Use the condition-combined Mann-Whitney test for each individual block. This method tests neural correlates with the visual stimulus side for each condition(choice, block side), and then combines them to get a single p-value for each neuron.  Neurons with p<0.05 are significantly correlated with the visual stimulus side.  Use PSTH for a sanity check.